requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyspark>=3.5.0

# FHIR support
fhir.resources>=7.0.2

# Data validation
great-expectations>=0.18.0
pydantic>=2.0.0

# Healthcare standards
python-hl7>=0.4.0

# Distributed computing
dask>=2023.12.0
ray>=2.8.0

# ML/AI tools
scikit-learn>=1.3.0
mlflow>=2.9.0

# Workflow orchestration
dagster>=1.5.0

# Monitoring & visualization
plotly>=5.0.0
dash>=2.0.0

# Database support
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Cloud storage
boto3>=1.26.0
google-cloud-storage>=2.0.0
azure-storage-blob>=12.0.0

# Utilities
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.65.0